services/cognitive/requirements.txt
# Core dependencies
pydantic==2.10.5
pydantic-settings==2.7.1
openai==1.59.6

# NLP
spacy==3.8.3

# Kafka
kafka-python==2.0.2

# API clients
requests==2.32.3

# Utilities
python-dotenv==1.0.1
